requirements.txt
disnake
mysql.connector
datetime
logging